Rename middleware folder
