+++ server.js
@@ -3,8 +3,8 @@
 const express = require('express')
 const app = express()
 
-const apiMiddleware = require('./middlewares/api')
-const redirectsMiddleware = require('./middlewares/redirects')
+const apiMiddleware = require('./middleware/api')
+const redirectsMiddleware = require('./middleware/redirects')
 
 const port = process.env.PORT || 3000
